Small fix: conversion warning
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4632 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/PhysicsLX56.cpp
@@ -407,7 +407,7 @@ public:
 			float friction = cClient->getGameLobby()->features[FT_WormFriction];
 			if(friction > 0) {
 				static const float wormSize = 5.0f;
-				static const float wormMass = (wormSize/2) * (wormSize/2) * PI;
+				static const float wormMass = (wormSize/2) * (wormSize/2) * (float)PI;
 				static const float wormDragCoeff = 0.1f; // Note: Never ever change this! (Or we have to make this configureable)
 				applyFriction(worm->velocity(), dt, wormSize, wormMass, wormDragCoeff, friction);
 			}
